Rename utils.rs to curves.rs
